+++ design-toolkit/components/src/components/accordion/accordion.stories.tsx
@@ -21,7 +21,7 @@ import type { Meta, StoryObj } from '@storybook/react';
 /**
  * The accordion will stretch to fill the entire width of its parent container.
  */
-const meta: Meta<typeof Accordion> = {
+const meta = {
   title: 'Components/Accordion',
   component: Accordion,
   args: {
@@ -39,11 +39,12 @@ const meta: Meta<typeof Accordion> = {
       subtitle: 'Content that can expand and collapse.',
     },
   },
-};
+} satisfies Meta<typeof Accordion>;
 
 export default meta;
+type Story = StoryObj<typeof meta>;
 
-export const Default: StoryObj<typeof Accordion> = {
+export const Default: Story = {
   render: ({ children, ...args }) => (
     <div className='w-[280px]'>
       <Accordion {...args}>
@@ -66,7 +67,7 @@ export const Default: StoryObj<typeof Accordion> = {
   ),
 };
 
-export const WithMenu: StoryObj<typeof Accordion> = {
+export const WithMenu: Story = {
   render: ({ children, ...args }) => (
     <div className='w-[280px]'>
       <Accordion {...args}>