+++ packages/bus/src/broadcast/index.test.ts
@@ -10,16 +10,21 @@
  * governing permissions and limitations under the License.
  */
 
+import { uuid } from '@accelint/core';
 import { afterEach, beforeEach, describe, expect, it, vi } from 'vitest';
 import {
   mockBroadcastChannel,
   resetMockBroadcastChannel,
 } from 'vitest-broadcast-channel-mock';
 import { Broadcast } from './index';
+import type { Payload } from './types';
 
 describe('broadcast', () => {
   beforeEach(() => {
     mockBroadcastChannel();
+
+    // Make sure to completely reset instance between tests
+    Broadcast.getInstance().destroy();
   });
 
   afterEach(() => {
@@ -27,7 +32,7 @@ describe('broadcast', () => {
   });
 
   it('on', () => {
-    const bus = Broadcast.getInstance();
+    const bus = Broadcast.getInstance<Payload<'test', string>>();
     const fn = vi.fn();
 
     bus.on('test', fn);
@@ -35,14 +40,14 @@ describe('broadcast', () => {
 
     expect(fn).toHaveBeenCalled();
     expect(fn).toHaveBeenCalledWith({
-      payload: 'test',
       type: 'test',
+      payload: 'test',
+      source: bus.id,
     });
-    expect(bus.getEvents()).toContain('test');
   });
 
   it('once', () => {
-    const bus = Broadcast.getInstance();
+    const bus = Broadcast.getInstance<Payload<'test', string>>();
     const fn = vi.fn();
 
     bus.once('test', fn);
@@ -52,14 +57,51 @@ describe('broadcast', () => {
 
     expect(fn).toHaveBeenCalledOnce();
     expect(fn).toHaveBeenCalledWith({
-      payload: 'test',
       type: 'test',
+      payload: 'test',
+      source: bus.id,
     });
-    expect(bus.getEvents()).not.toContain('test');
+  });
+
+  it('on & once', () => {
+    const bus = Broadcast.getInstance<Payload<'test', string>>();
+    const on = vi.fn();
+    const once = vi.fn();
+
+    bus.on('test', on);
+    bus.once('test', once);
+
+    bus.emit('test', 'A');
+    bus.emit('test', 'B');
+    bus.emit('test', 'C');
+
+    expect(on).toHaveBeenCalledTimes(3);
+    expect(on).toHaveBeenNthCalledWith(1, {
+      type: 'test',
+      payload: 'A',
+      source: bus.id,
+    });
+    expect(on).toHaveBeenNthCalledWith(2, {
+      type: 'test',
+      payload: 'B',
+      source: bus.id,
+    });
+    expect(on).toHaveBeenNthCalledWith(3, {
+      type: 'test',
+      payload: 'C',
+      source: bus.id,
+    });
+
+    expect(once).toHaveBeenCalledOnce();
+    expect(once).toHaveBeenCalledWith({
+      type: 'test',
+      payload: 'A',
+      source: bus.id,
+    });
   });
 
   it('off', () => {
-    const bus = Broadcast.getInstance();
+    const bus = Broadcast.getInstance<Payload<'test', string>>();
     const fn = vi.fn();
 
     bus.on('test', fn);
@@ -67,11 +109,10 @@ describe('broadcast', () => {
     bus.emit('test', 'test');
 
     expect(fn).not.toHaveBeenCalled();
-    expect(bus.getEvents()).toContain('test');
   });
 
   it('destroy', () => {
-    const bus = Broadcast.getInstance();
+    const bus = Broadcast.getInstance<Payload<'test', string>>();
     const fn = vi.fn();
 
     bus.on('test', fn);
@@ -79,6 +120,80 @@ describe('broadcast', () => {
     bus.emit('test', 'test');
 
     expect(fn).not.toHaveBeenCalled();
-    expect(bus.getEvents()).toEqual([]);
+  });
+
+  it('should deliver to all', () => {
+    const bus = Broadcast.getInstance<Payload<'test', string>>();
+    const fn = vi.fn();
+
+    bus.on('test', fn);
+    bus.emit('test', 'all', { target: 'all' });
+
+    expect(fn).toHaveBeenCalledTimes(1);
+    expect(fn).toHaveBeenCalledWith({
+      type: 'test',
+      payload: 'all',
+      source: bus.id,
+    });
+  });
+
+  it('should deliver to self', () => {
+    const bus = Broadcast.getInstance<Payload<'test', string>>();
+    const fn = vi.fn();
+
+    bus.on('test', fn);
+    bus.emit('test', 'self', { target: 'self' });
+
+    expect(fn).toHaveBeenCalledTimes(1);
+    expect(fn).toHaveBeenCalledWith({
+      type: 'test',
+      payload: 'self',
+      source: bus.id,
+      target: bus.id,
+    });
+  });
+
+  it('should deliver to others', () => {
+    const bus = Broadcast.getInstance<Payload<'test', string>>();
+    const fn = vi.fn();
+
+    bus.on('test', fn);
+    bus.emit('test', 'echo', { target: 'others' });
+
+    expect(fn).not.toHaveBeenCalled();
+  });
+
+  it('should default to all as target audience', () => {
+    const bus = Broadcast.getInstance<Payload<'test', string>>();
+    const fn = vi.fn();
+
+    bus.on('test', fn);
+    bus.emit('test', 'default');
+
+    expect(fn).toHaveBeenCalledTimes(1);
+    expect(fn).toHaveBeenCalledWith({
+      type: 'test',
+      payload: 'default',
+      source: bus.id,
+    });
+  });
+
+  it('should deliver to specific target', () => {
+    const bus = Broadcast.getInstance<Payload<'test', string>>();
+    const fn = vi.fn();
+    const target = uuid();
+
+    bus.on('test', fn);
+    bus.emit('test', 'test', { target });
+
+    expect(fn).not.toHaveBeenCalled();
+
+    // @ts-expect-error Accessing protected property
+    expect(bus.channel.postMessage).toHaveBeenCalledWith({
+      type: 'test',
+      payload: 'test',
+      source: bus.id,
+      target,
+    });
   });
 });